Set function call span attributes on all spans instead of just the unknown spans.
