+++ src/core/trulens/core/otel/instrument.py
@@ -11,10 +11,12 @@ from opentelemetry.baggage import remove_baggage
 from opentelemetry.baggage import set_baggage
 import opentelemetry.context as context_api
 from opentelemetry.trace.span import Span
-from opentelemetry.util.types import AttributeValue
 from trulens.core import app as core_app
 from trulens.experimental.otel_tracing.core.session import TRULENS_SERVICE_NAME
 from trulens.experimental.otel_tracing.core.span import Attributes
+from trulens.experimental.otel_tracing.core.span import (
+    set_function_call_attributes,
+)
 from trulens.experimental.otel_tracing.core.span import (
     set_general_span_attributes,
 )
@@ -103,6 +105,8 @@ def _set_span_attributes(
         args_with_self_possibly = (instance,) + args
     else:
         args_with_self_possibly = args
+    # Set function call attributes.
+    set_function_call_attributes(span, ret, func_exception, all_kwargs)
     # Combine the attributes with the full_scoped_attributes.
     resolved_attributes = _resolve_attributes(
         attributes,
@@ -126,33 +130,13 @@ def _set_span_attributes(
         **resolved_attributes,
         **resolved_full_scoped_attributes,
     }
-    if span_type == SpanAttributes.SpanType.UNKNOWN and not all_attributes:
-        all_attributes = {
-            f"{BASE_SCOPE}.{span_type.value}.{k}": v
-            for k, v in all_kwargs.items()
-        }
-        all_attributes[f"{BASE_SCOPE}.{span_type.value}.return"] = ret
-    all_attributes = {
-        k: _convert_to_valid_span_attribute_type(v)
-        for k, v in all_attributes.items()
-    }
-    # Set the user-provided attributes.
-    set_user_defined_attributes(
-        span,
-        span_type=span_type,
-        attributes=all_attributes,
-    )
-
-
-def _convert_to_valid_span_attribute_type(val: Any) -> AttributeValue:
-    if isinstance(val, (bool, int, float, str)):
-        return val
-    if isinstance(val, (list, tuple)):
-        for curr_type in [bool, int, float, str]:
-            if all([isinstance(curr, curr_type) for curr in val]):
-                return val
-        return [str(curr) for curr in val]
-    return str(val)
+    if all_attributes:
+        # Set the user-provided attributes.
+        set_user_defined_attributes(
+            span,
+            span_type=span_type,
+            attributes=all_attributes,
+        )
 
 
 def instrument(
