+++ src/core/trulens/experimental/otel_tracing/core/span.py
@@ -4,10 +4,12 @@ This file contains utility functions specific to certain span types.
 
 from inspect import signature
 import logging
+import traceback
 from typing import Any, Callable, Dict, List, Optional, Union
 
 from opentelemetry.baggage import get_baggage
 from opentelemetry.trace.span import Span
+from opentelemetry.util.types import AttributeValue
 from trulens.core.utils import signature as signature_utils
 from trulens.otel.semconv.trace import BASE_SCOPE
 from trulens.otel.semconv.trace import SpanAttributes
@@ -74,6 +76,26 @@ def validate_selector_name(attributes: Dict[str, Any]) -> Dict[str, Any]:
     return result
 
 
+def _convert_to_valid_span_attribute_type(val: Any) -> AttributeValue:
+    if isinstance(val, (bool, int, float, str)):
+        return val
+    if isinstance(val, (list, tuple)):
+        for curr_type in [bool, int, float, str]:
+            if all([isinstance(curr, curr_type) for curr in val]):
+                return val
+        return [str(curr) for curr in val]
+    return str(val)
+
+
+def set_span_attribute_safely(
+    span: Span,
+    key: str,
+    value: Any,
+) -> None:
+    if value is not None:
+        span.set_attribute(key, _convert_to_valid_span_attribute_type(value))
+
+
 def validate_attributes(attributes: Dict[str, Any]) -> Dict[str, Any]:
     """
     Utility function to validate span attributes based on the span type.
@@ -92,7 +114,7 @@ def validate_attributes(attributes: Dict[str, Any]) -> Dict[str, Any]:
 
 def set_general_span_attributes(
     span: Span, /, span_type: SpanAttributes.SpanType
-) -> Span:
+) -> None:
     span.set_attribute(SpanAttributes.SPAN_TYPE, span_type)
 
     span.set_attribute(
@@ -117,7 +139,20 @@ def set_general_span_attributes(
     if input_id_baggage:
         span.set_attribute(SpanAttributes.INPUT_ID, str(input_id_baggage))
 
-    return span
+
+def set_function_call_attributes(
+    span: Span,
+    ret: Any,
+    func_exception: Optional[Exception],
+    all_kwargs: Dict[str, Any],
+) -> None:
+    set_span_attribute_safely(span, SpanAttributes.CALL.RETURN, ret)
+    set_span_attribute_safely(span, SpanAttributes.CALL.ERROR, func_exception)
+    set_span_attribute_safely(
+        span, SpanAttributes.CALL.STACK, "\n".join(traceback.format_stack())
+    )
+    for k, v in all_kwargs.items():
+        set_span_attribute_safely(span, f"{SpanAttributes.CALL.KWARGS}.{k}", v)
 
 
 def set_user_defined_attributes(
@@ -129,13 +164,13 @@ def set_user_defined_attributes(
     final_attributes = validate_attributes(attributes)
 
     for key, value in final_attributes.items():
-        span.set_attribute(key, value)
-
+        set_span_attribute_safely(span, key, value)
         if (
             key != SpanAttributes.SELECTOR_NAME_KEY
             and SpanAttributes.SELECTOR_NAME_KEY in final_attributes
         ):
-            span.set_attribute(
+            set_span_attribute_safely(
+                span,
                 f"{BASE_SCOPE}.{final_attributes[SpanAttributes.SELECTOR_NAME_KEY]}.{key}",
                 value,
             )
@@ -161,15 +196,18 @@ def set_main_span_attributes(
     ret: Any,
     exception: Optional[Exception],
 ) -> None:
-    span.set_attribute(
-        SpanAttributes.MAIN.MAIN_INPUT, get_main_input(func, args, kwargs)
+    set_span_attribute_safely(
+        span, SpanAttributes.MAIN.MAIN_INPUT, get_main_input(func, args, kwargs)
     )
 
     if exception:
-        span.set_attribute(SpanAttributes.MAIN.MAIN_ERROR, str(exception))
+        set_span_attribute_safely(
+            span, SpanAttributes.MAIN.MAIN_ERROR, str(exception)
+        )
 
     if ret is not None:
-        span.set_attribute(
+        set_span_attribute_safely(
+            span,
             SpanAttributes.MAIN.MAIN_OUTPUT,
             signature_utils.main_output(func, ret),
         )
